refactor(orchestra): extract the dev debug panel into @novus/orchestra
Move dev/ (Orchestra tools + grid/minimap/stats/cmdo + Theatre.js
integration) into packages/orchestra as the @novus/orchestra workspace
package. webgl/ now consumes the Theatre sheet API via
@novus/orchestra/theatre instead of relative dev/ paths. @base-ui/react,
@theatre/studio, and stats-gl move from root deps to the package; root
keeps @theatre/core (webgl uses it directly).
