+++ package.json
@@ -19,8 +19,8 @@
     "clean": "rm -rf build node_modules/.cache .react-router"
   },
   "dependencies": {
-    "@base-ui/react": "^1.3.0",
     "@novus/font-optimizer": "workspace:*",
+    "@novus/orchestra": "workspace:*",
     "@novus/password-protection": "workspace:*",
     "@novus/static-i18n": "workspace:*",
     "@novus/transitions": "workspace:*",
@@ -37,7 +37,6 @@
     "@sanity/visual-editing": "^5.3.3",
     "@t3-oss/env-core": "^0.13.11",
     "@theatre/core": "^0.7.2",
-    "@theatre/studio": "^0.7.2",
     "@vercel/react-router": "^1.2.6",
     "animejs": "^4.3.6",
     "clsx": "^2.1.1",
@@ -49,7 +48,6 @@
     "react": "19.2.4",
     "react-dom": "19.2.4",
     "react-router": "^7.14.0",
-    "stats-gl": "^4.1.0",
     "tempus": "1.0.0-dev.17",
     "three": "^0.183.2",
     "tunnel-rat": "^0.1.2",
