+++ packages/orchestra/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@novus/orchestra",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./index.tsx",
+    "./theatre": "./theatre/index.tsx",
+    "./theatre/hooks/use-theatre": "./theatre/hooks/use-theatre.ts"
+  },
+  "dependencies": {
+    "@base-ui/react": "^1.3.0",
+    "@theatre/core": "^0.7.2",
+    "@theatre/studio": "^0.7.2",
+    "clsx": "^2.1.1",
+    "hamo": "1.0.0-dev.13",
+    "stats-gl": "^4.1.0",
+    "tempus": "1.0.0-dev.17",
+    "three": "^0.183.2",
+    "zustand": "^5.0.12"
+  },
+  "peerDependencies": {
+    "react": ">=19"
+  }
+}
