+++ packages/orchestra/theatre/studio/index.tsx
