+++ webgl/utils/fluid/index.tsx
@@ -2,8 +2,8 @@ import { useFrame, useThree } from "@react-three/fiber";
 import theatreCore from "@theatre/core";
 const { types } = theatreCore;
 import { useEffect, useRef } from "react";
-import { useCurrentSheet } from "../../../dev/theatre";
-import { useTheatre } from "../../../dev/theatre/hooks/use-theatre";
+import { useCurrentSheet } from "@novus/orchestra/theatre";
+import { useTheatre } from "@novus/orchestra/theatre/hooks/use-theatre";
 import { Fluid } from "./fluid-sim";
 
 export function useFluidSim() {
